refactor: extract MPV embedding component into standalone iced-mpv crate
