+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "crates/media-sort-core",
     "crates/media-sort-backend",
+    "crates/iced-mpv",
     "crates/iced-automation",
     "crates/iced-automation-macros",
     "crates/media-sort-gui", "crates/benchmarks",
