+++ crates/benchmarks/Cargo.toml
@@ -14,7 +14,7 @@ image.workspace = true
 fast_image_resize.workspace = true
 zune-jpeg = "0.5"
 zune-core = "0.5"
-libmpv-sys = "3.1.0"
+iced-mpv = { path = "../iced-mpv", default-features = false }
 turbojpeg = "1.5"
 kamadak-exif.workspace = true
 webp = "0.3"
