+++ crates/iced-mpv/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "iced-mpv"
+version.workspace = true
+edition.workspace = true
+license = "MIT"
+repository = "https://github.com/Lolle2000la/Image-Sort"
+
+[features]
+default = ["ui"]
+ui = ["dep:lucide-icons"]
+
+[dependencies]
+libmpv-sys = "3.1.0"
+iced = { workspace = true, features = ["wgpu", "tokio", "advanced"] }
+iced_wgpu = "0.14"
+bytemuck = "1"
+tokio = { workspace = true, features = ["rt", "rt-multi-thread", "sync", "macros", "time"] }
+tracing = { workspace = true }
+rayon = { workspace = true }
+kamadak-exif = { workspace = true }
+mp4ameta = { workspace = true }
+lucide-icons = { version = "1.21", optional = true }
