+++ crates/iced-mpv/src/widget.rs
@@ -1,0 +1,6 @@
+#[cfg(feature = "ui")]
+pub mod controls;
+#[cfg(feature = "ui")]
+pub mod player;
+
+pub mod shader;
