+++ crates/media-sort-backend/Cargo.toml
@@ -19,7 +19,7 @@ mp4ameta.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 tokio.workspace = true
-libmpv-sys = "3.1.0"
+iced-mpv = { path = "../iced-mpv", default-features = false }
 chrono = "0.4"
 fast_image_resize.workspace = true
 rayon.workspace = true
