+++ crates/media-sort-backend/src/media.rs
@@ -5,5 +5,4 @@ pub mod audio_decoder;
 pub mod ffmpeg_pipe;
 mod format_pipeline;
 pub mod image_decoder;
-pub mod mpv_context;
 pub mod thumbnail;
