+++ crates/media-sort-gui/Cargo.toml
@@ -19,6 +19,7 @@ demo = [
 [dependencies]
 media-sort-core = { path = "../media-sort-core" }
 media-sort-backend = { path = "../media-sort-backend" }
+iced-mpv = { path = "../iced-mpv" }
 iced-automation = { path = "../iced-automation", optional = true }
 iced-automation-macros = { path = "../iced-automation-macros" }
 iced.workspace = true
