+++ crates/media-sort-gui/src/message.rs
@@ -161,8 +161,11 @@ pub enum MediaMessage {
 
 #[derive(Debug, Clone)]
 pub enum VideoMessage {
-    PlayerReady(tokio::sync::mpsc::Sender<media_sort_backend::media::mpv_context::VideoCommand>),
-    Event(media_sort_backend::media::mpv_context::VideoEvent),
+    Player(iced_mpv::PlayerMessage),
+    #[allow(dead_code)]
+    PlayerReady(tokio::sync::mpsc::Sender<iced_mpv::VideoCommand>),
+    #[allow(dead_code)]
+    Event(iced_mpv::VideoEvent),
     Seek(f64),
     Volume(f64),
     Mute,
