+++ crates/media-sort-gui/src/subscriptions.rs
@@ -1,4 +1,3 @@
 pub mod keyboard;
 pub mod prefetch;
 pub mod thumbnail_tracker;
-pub mod video_player;
