+++ crates/media-sort-gui/src/widgets.rs
@@ -1,7 +1,5 @@
 pub mod create_folder_modal;
 pub mod folder_icon;
-pub mod media_controls;
 pub mod rename_modal;
 pub mod video_canvas;
 pub mod video_player;
-pub mod video_shader;
