+++ crates/media-sort-gui/src/widgets/video_canvas.rs
@@ -2,10 +2,10 @@ use std::path::PathBuf;
 
 use iced::widget::{column, container, text};
 use iced::{Color, Element, Font, Length};
+use iced_mpv::{MediaControlMessage, media_controls_view};
 
 use crate::message::{MediaMessage, Message};
 use crate::state::AppState;
-use crate::widgets::media_controls;
 
 pub fn audio_controls_view(
     _path: PathBuf,
@@ -35,7 +35,7 @@ pub fn audio_controls_view(
         .into()
     };
 
-    let controls = media_controls::media_controls_view(
+    let controls = media_controls_view(
         state.audio.position,
         state.audio.duration,
         state.audio.volume,
@@ -43,23 +43,15 @@ pub fn audio_controls_view(
         playing,
     )
     .map(|msg| match msg {
-        media_controls::MediaControlMessage::PlayPause => {
-            Message::Media(MediaMessage::AudioPlayPause)
-        }
-        media_controls::MediaControlMessage::Stop => Message::Media(MediaMessage::StopAudio),
-        media_controls::MediaControlMessage::Seek(v) => Message::Media(MediaMessage::AudioSeek(v)),
-        media_controls::MediaControlMessage::SetVolume(v) => {
-            Message::Media(MediaMessage::AudioSetVolume(v))
-        }
-        media_controls::MediaControlMessage::ToggleMute => {
-            Message::Media(MediaMessage::AudioToggleMute)
-        }
+        MediaControlMessage::PlayPause => Message::Media(MediaMessage::AudioPlayPause),
+        MediaControlMessage::Stop => Message::Media(MediaMessage::StopAudio),
+        MediaControlMessage::Seek(v) => Message::Media(MediaMessage::AudioSeek(v)),
+        MediaControlMessage::SetVolume(v) => Message::Media(MediaMessage::AudioSetVolume(v)),
+        MediaControlMessage::ToggleMute => Message::Media(MediaMessage::AudioToggleMute),
     });
 
     let content = column![
         container(audio_content)
-            .center_x(Length::Fill)
-            .center_y(Length::Fill)
             .width(Length::Fill)
             .height(Length::Fill),
         controls
