+++ crates/media-sort-gui/src/widgets/video_player.rs
@@ -5,59 +5,28 @@ use iced::{Alignment, Element, Font, Length};
 
 use crate::message::{Message, VideoMessage};
 use crate::state::AppState;
-use crate::widgets::media_controls;
 
 pub fn video_player<'a>(
     path: PathBuf,
     state: &AppState,
     thumb_handle: Option<iced::widget::image::Handle>,
 ) -> Element<'a, Message> {
-    let video_content: Element<'_, Message> = if state.video.rgba.is_some() {
-        crate::widgets::video_shader::video_shader_view(
-            state.video.width,
-            state.video.height,
-            state.video.rotation,
-            state.video.rgba.clone(),
-        )
-    } else if let Some(handle) = thumb_handle {
-        iced::widget::image(handle)
-            .width(Length::Fill)
-            .height(Length::Fill)
-            .into()
-    } else {
-        placeholder(path, &state.l10n)
-    };
-
-    let display_position = state.video.seek_position.unwrap_or(state.video.position);
-
-    let controls_row = media_controls::media_controls_view(
-        display_position,
-        state.video.duration,
-        state.video.volume,
-        state.video.muted,
-        !state.video.paused,
+    iced_mpv::video_player_view(
+        path.clone(),
+        &state.video,
+        thumb_handle,
+        Some(placeholder(path, &state.l10n)),
+        |action| match action {
+            iced_mpv::VideoAction::PlayPause => Message::Video(VideoMessage::PlayPause),
+            iced_mpv::VideoAction::Stop => Message::Video(VideoMessage::Stop),
+            iced_mpv::VideoAction::Seek(v) => Message::Video(VideoMessage::Seek(v)),
+            iced_mpv::VideoAction::SetVolume(v) => Message::Video(VideoMessage::Volume(v)),
+            iced_mpv::VideoAction::ToggleMute => Message::Video(VideoMessage::Mute),
+            iced_mpv::VideoAction::PlayExternally(p) => {
+                Message::Video(VideoMessage::PlayExternally(p))
+            }
+        },
     )
-    .map(|msg| match msg {
-        media_controls::MediaControlMessage::PlayPause => Message::Video(VideoMessage::PlayPause),
-        media_controls::MediaControlMessage::Stop => Message::Video(VideoMessage::Stop),
-        media_controls::MediaControlMessage::Seek(v) => Message::Video(VideoMessage::Seek(v)),
-        media_controls::MediaControlMessage::SetVolume(v) => {
-            Message::Video(VideoMessage::Volume(v))
-        }
-        media_controls::MediaControlMessage::ToggleMute => Message::Video(VideoMessage::Mute),
-    });
-
-    column![
-        container(video_content)
-            .width(Length::Fill)
-            .height(Length::Fill)
-            .center_x(Length::Fill)
-            .center_y(Length::Fill),
-        controls_row
-    ]
-    .width(Length::Fill)
-    .height(Length::Fill)
-    .into()
 }
 
 fn placeholder(
